Added AddNodePage
The AddNodePage will be a page containing multiple steps to add a node.
Currently step one is added which is opening the Bluetooth settings and connecting to the node.
There is a "Next Step" and "Previous Step" button to navigate through the steps
The AddNodePage is accessed by the FloatingActionButton on the NodeSummaryPage
To open the Bluetooth settings via a button the open_settings
package was used