Added Bluetooth scan and step 2
Moved the building of a step and, functions of a step, to it's own separate file
The first page now exists as AddNodeStep1
This has been done to minimize clutter as every step in a switch case would have been a nightmare
Bluetooth package flutter_blue 0.6.2 has been added. Latest version causes crashes
Added AddNodeStep2. It will perform a 5 second long scan for devices and display the FarmLab devices
A single device can be selected
This device will become the device you get connected to (not implemented)