Added token management and autologin
Tokens are now kept by the SharedPreferences
This means they are kept inbetween sessions
The saved sata can be accesed by the store, save and clear functions
On startup there will be checked if there is a saved token.
If there is you will be logged in automatically
If not the token will be saved when you log in manually
NodeSummaryPage has a logout button that clears the tokens and returns to LoginPage
Refixed issue that prevented helpers from returning errors
Refactored NetworkHelper
Added refresh functionality to NetworkHelper.
It will automatically request a new token when it expires
Removed refresh from AuthenticationHelper as its functionality has been added to NetworkHelper instead